pyproject.toml
[project]
name = "mtn-exec-copilot"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.5",
    "azure-ai-projects>=2.1.0",
    "azure-ai-voicelive[aiohttp]>=1.0.0",
    "azure-core>=1.41.0",
    "azure-identity>=1.26.0b2",
    "fastapi>=0.104.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.24.0",
]

[tool.uv]
prerelease = "allow"
package = true

[project.scripts]
mtn-exec-copilot = "backend.main:run"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend", "backend.api", "backend.voice"]
